format efixloader, change main source name
