+++ efiXloader/.clang-format
@@ -2,4 +2,4 @@
 Language: Cpp
 BasedOnStyle: LLVM
 ColumnLimit: 90
-IndentWidth: 4
+IndentWidth: 2
