+++ efiXloader/pe_manager.cpp
@@ -1,6 +1,6 @@
 /*
  * efiXloader
- * Copyright (C) 2020-2023 Binarly
+ * Copyright (C) 2020-2024 Binarly
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -21,13 +21,13 @@
 #include "pe_manager.h"
 
 void efiloader::PeManager::process(linput_t *li, std::basic_string<char> fname, int ord) {
-    efiloader::PE pe(li, fname, &pe_base, &pe_sel_base, ord, machine_type);
-    if (pe.good() && pe.is_p32_plus()) {
-        msg("[efiXloader] PE detected\n");
-        pe.process();
-    } else if (pe.is_p32()) {
-        msg("[efiXloader] this loader is not ready for PE32\n");
-    } else {
-        warning("[efiXloader] not PE\n");
-    }
+  efiloader::PE pe(li, fname, &pe_base, &pe_sel_base, ord, machine_type);
+  if (pe.good() && pe.is_p32_plus()) {
+    msg("[efiXloader] PE detected\n");
+    pe.process();
+  } else if (pe.is_p32()) {
+    msg("[efiXloader] this loader is not ready for PE32\n");
+  } else {
+    warning("[efiXloader] not PE\n");
+  }
 }
