+++ efiXloader/uefitool.h
@@ -1,6 +1,6 @@
 /*
  * efiXloader
- * Copyright (C) 2020-2023 Binarly
+ * Copyright (C) 2020-2024 Binarly
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -53,96 +53,96 @@
 using namespace nlohmann;
 
 enum FILE_SECTION_TYPE {
-    PE_DEPENDENCY_SECTION = 0,
-    PE_TE_IMAGE_SECTION = 1,
-    UI_SECTION = 2,
-    VERSION_SECTION = 3
+  PE_DEPENDENCY_SECTION = 0,
+  PE_TE_IMAGE_SECTION = 1,
+  UI_SECTION = 2,
+  VERSION_SECTION = 3
 };
 
 namespace efiloader {
 class File {
-  public:
-    File() {}
-    void set_data(char *data_in, uint32_t size_in) {
-        qname.qclear();
-        bytes.resize(size_in);
-        memcpy(&bytes[0], data_in, size_in);
-    };
-    void write() {
-        qstring idb_path(get_path(PATH_TYPE_IDB));
-        qstring images_path = idb_path + qstring(".efiloader");
+public:
+  File() {}
+  void set_data(char *data_in, uint32_t size_in) {
+    qname.qclear();
+    bytes.resize(size_in);
+    memcpy(&bytes[0], data_in, size_in);
+  };
+  void write() {
+    qstring idb_path(get_path(PATH_TYPE_IDB));
+    qstring images_path = idb_path + qstring(".efiloader");
 #ifdef WIN32
-        _mkdir(images_path.c_str());
+    _mkdir(images_path.c_str());
 #else
-        mkdir(images_path.c_str(), S_IRWXU | S_IRWXG | S_IROTH | S_IXOTH);
+    mkdir(images_path.c_str(), S_IRWXU | S_IRWXG | S_IROTH | S_IXOTH);
 #endif
-        if (!qname.empty()) {
-            qstring image_path = images_path + qstring("/") + qstring(qname.c_str());
-            std::ofstream file;
-            file.open(image_path.c_str(), std::ios::out | std::ios::binary);
-            file.write(ubytes.constData(), ubytes.size());
-            file.close();
-            dump_name.swap(image_path);
-        }
+    if (!qname.empty()) {
+      qstring image_path = images_path + qstring("/") + qstring(qname.c_str());
+      std::ofstream file;
+      file.open(image_path.c_str(), std::ios::out | std::ios::binary);
+      file.write(ubytes.constData(), ubytes.size());
+      file.close();
+      dump_name.swap(image_path);
     }
-    void print();
-    UByteArray ubytes;
-    UByteArray uname;
-    bytevec_t bytes;
-    char *data = NULL;
-    uint32_t size = 0;
-    std::string name_utf8;
-    std::string name_utf16;
-    qstring qname;
-    qstring dump_name;
-    bool is_pe = false;
-    bool is_te = false;
-    bool has_ui = false;
+  }
+  void print();
+  UByteArray ubytes;
+  UByteArray uname;
+  bytevec_t bytes;
+  char *data = NULL;
+  uint32_t size = 0;
+  std::string name_utf8;
+  std::string name_utf16;
+  qstring qname;
+  qstring dump_name;
+  bool is_pe = false;
+  bool is_te = false;
+  bool has_ui = false;
 };
 
 class Uefitool {
-  public:
-    Uefitool(bytevec_t &data) {
-        buffer = (const char *)&data[0];
-        buffer_size = data.size();
-        UByteArray ubuffer(buffer, buffer_size);
-        FfsParser ffs(&model);
-        if (ffs.parse(ubuffer)) {
-            loader_failure("failed to parse data via UEFITool");
-        }
-        messages = ffs.getMessages();
+public:
+  Uefitool(bytevec_t &data) {
+    buffer = (const char *)&data[0];
+    buffer_size = data.size();
+    UByteArray ubuffer(buffer, buffer_size);
+    FfsParser ffs(&model);
+    if (ffs.parse(ubuffer)) {
+      loader_failure("failed to parse data via UEFITool");
     }
-    ~Uefitool() {};
-    void show_messages();
-    bool messages_occurs() { return !messages.empty(); };
-    void dump();
-    void dump(const UModelIndex &index);
-    void dump(const UModelIndex &index, uint8_t el_type, File *pe_file);
-    void handle_raw_section(const UModelIndex &index);
-    bool is_pe_index(const UModelIndex &index) { return model.rowCount(index) == 4; };
-    bool is_file_index(const UModelIndex &index) {
-        return model.type(index) == Types::File;
-    };
-    void get_unique_name(qstring &image_name);
-    void get_image_guid(qstring &image_guid, UModelIndex index);
-    std::vector<std::string> parseDepexSectionBody(const UModelIndex &index,
-                                                   UString &parsed);
-    std::vector<std::string> parseAprioriRawSection(const UModelIndex &index);
-    void get_deps(UModelIndex index, std::string key);
-    void get_apriori(UModelIndex index, std::string key);
-    void dump_jsons(); // dump JSON with DEPEX and GUIDs information for each image
-    json all_deps;     // DEPEX information for each image
-    json images_guids; // matching the modules to the parent's GUIDs
-    TreeModel model;
-    const char *buffer;
-    uint32_t buffer_size;
-    std::vector<std::pair<UString, UModelIndex>> messages;
-    std::set<qstring> unique_names;
-    std::vector<efiloader::File *> files;
-    USTATUS err;
-    void set_machine_type(UByteArray pe_body);
-    uint16_t machine_type = 0xffff;
-    bool machine_type_detected = false;
+    messages = ffs.getMessages();
+  }
+  ~Uefitool() {};
+  void show_messages();
+  bool messages_occurs() { return !messages.empty(); };
+  void dump();
+  void dump(const UModelIndex &index);
+  void dump(const UModelIndex &index, uint8_t el_type, File *pe_file);
+  void handle_raw_section(const UModelIndex &index);
+  bool is_pe_index(const UModelIndex &index) { return model.rowCount(index) == 4; };
+  bool is_file_index(const UModelIndex &index) {
+    return model.type(index) == Types::File;
+  };
+  void get_unique_name(qstring &image_name);
+  void get_image_guid(qstring &image_guid, UModelIndex index);
+  std::vector<std::string> parseDepexSectionBody(const UModelIndex &index,
+                                                 UString &parsed);
+  std::vector<std::string> parseAprioriRawSection(const UModelIndex &index);
+  void get_deps(UModelIndex index, std::string key);
+  void get_apriori(UModelIndex index, std::string key);
+  void dump_jsons(); // dump JSON with DEPEX and GUIDs information for each image
+  json all_deps;     // DEPEX information for each image
+  json images_guids; // matching the modules to the parent's GUIDs
+  TreeModel model;
+  const char *buffer;
+  uint32_t buffer_size;
+  std::vector<std::pair<UString, UModelIndex>> messages;
+  std::set<qstring> unique_names;
+  std::vector<efiloader::File *> files;
+  USTATUS err;
+  void set_machine_type(UByteArray pe_body);
+  uint16_t machine_type = 0xffff;
+  bool machine_type_detected = false;
 };
 } // namespace efiloader
